requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
streamlit>=1.35.0
pandas>=2.0.0
plotly>=5.20.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
